Split resolver module metadata tests
